Addressing state structure comments
