+++ lib/main.dart
@@ -5,9 +5,9 @@ library;
 
 import 'package:ddr_md/components/bpm_page.dart';
 import 'package:ddr_md/components/song/song_page.dart';
+import 'package:ddr_md/models/bpm_model.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:ddr_md/constants.dart' as constants;
 
 void main() {
   runApp(const App());
@@ -19,7 +19,7 @@ class App extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ChangeNotifierProvider(
-      create: (context) => AppState(),
+      create: (context) => BpmState(),
       child: MaterialApp(
         debugShowCheckedModeBanner: false,
         title: 'ddr_bpm',
@@ -37,32 +37,6 @@ class App extends StatelessWidget {
       ),
     );
   }
-}
-
-class AppState extends ChangeNotifier {
-  int bpm = constants.chosenBpm; // BPM init
-
-  // Set BPM to new input
-  void setBpm(newBpm) {
-    if (newBpm == "") return;
-    bpm = int.parse(newBpm);
-    notifyListeners();
-  }
-
-  // Generate list of mods programmatically
-  List<double> generateMods() {
-    double start = 0.25, end = 4.25;
-    // Generate numbers from [0 to 4 in .25 inc.]
-    var list = [for (var i = start; i < end; i += .25) i];
-    start = 4.5;
-    end = 8.5;
-    // Generate numbers from [4 to 8 in .5 inc.]
-    [for (var i = start; i < end; i += .5) list.add(i)];
-    mods = list;
-    return list;
-  }
-
-  late var mods = generateMods();
 }
 
 class Navbar extends StatefulWidget {
